Issue 144: Move entities into "Domain/Model" Introduce the "Domain" Layer
