+++ back/config/packages/doctrine.yaml
@@ -23,5 +23,5 @@ doctrine:
                 is_bundle: false
                 type: yml
                 dir: '%kernel.project_dir%/config/doctrine'
-                prefix: 'App\Entity'
+                prefix: 'App\Domain\Model'
                 alias: App
