+++ back/src/Domain/Model/BlogPost.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace App\Entity;
+namespace App\Domain\Model;
 
 use Ramsey\Uuid\UuidInterface;
