+++ back/tests/spec/Domain/Model/BlogPostSpec.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace spec\App\Entity;
+namespace spec\App\Domain\Model;
 
 use PhpSpec\ObjectBehavior;
 use Ramsey\Uuid\Uuid;
